refactor: apply maven package changes
