+++ src/main/java/org/jnbt/Tag.java
@@ -1,0 +1,111 @@
+package org.jnbt;
+
+//@formatter:off
+
+/*
+ * JNBT License
+ * 
+ * Copyright (c) 2010 Graham Edgecombe
+ * All rights reserved.
+ * 
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 
+ *     * Redistributions of source code must retain the above copyright notice,
+ *       this list of conditions and the following disclaimer.
+ *       
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *       
+ *     * Neither the name of the JNBT team nor the names of its
+ *       contributors may be used to endorse or promote products derived from
+ *       this software without specific prior written permission.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE. 
+ */
+
+//@formatter:on
+
+/**
+ * Represents a single NBT tag.
+ * 
+ * @author Graham Edgecombe
+ * 
+ */
+public abstract class Tag {
+	
+	/**
+	 * The name of this tag.
+	 */
+	private final String name;
+	
+	/**
+	 * Creates the tag with the specified name.
+	 * 
+	 * @param name
+	 *            The name.
+	 */
+	public Tag(final String name) {
+	
+		this.name = name;
+	}
+	
+	/**
+	 * Gets the name of this tag.
+	 * 
+	 * @return The name of this tag.
+	 */
+	public final String getName() {
+	
+		return name;
+	}
+	
+	/**
+	 * Gets the value of this tag.
+	 * 
+	 * @return The value of this tag.
+	 */
+	public abstract Object getValue();
+	
+	/*
+	 * (non-Javadoc)
+	 * @see java.lang.Object#hashCode()
+	 */
+	@Override
+	public int hashCode() {
+	
+		final int prime = 31;
+		int result = 1;
+		result = (prime * result) + ((name == null) ? 0 : name.hashCode());
+		return result;
+	}
+	
+	/*
+	 * (non-Javadoc)
+	 * @see java.lang.Object#equals(java.lang.Object)
+	 */
+	@Override
+	public boolean equals(final Object obj) {
+	
+		if (this == obj) { return true; }
+		if (obj == null) { return false; }
+		if (!(obj instanceof Tag)) { return false; }
+		final Tag other = (Tag) obj;
+		if (name == null) {
+			if (other.name != null) { return false; }
+		} else if (!name.equals(other.name)) { return false; }
+		return true;
+	}
+	
+}
